ExpertSearch component.
